+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/UrnBlock.java
@@ -37,6 +37,7 @@ import net.minecraft.world.level.*;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.EntityBlock;
 import net.minecraft.world.level.block.FallingBlock;
+import net.minecraft.world.level.block.SandBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
@@ -108,23 +109,13 @@ public class UrnBlock extends FallingBlock implements EntityBlock {
         return false;
     }
 
-    @Override
-    public boolean canSurvive(BlockState state, LevelReader pLevel, BlockPos pos) {
-        return canSupportCenter(pLevel, pos.below(), Direction.UP);
-    }
-
     //called when a neighbor is placed
     @Override
     public BlockState updateShape(BlockState stateIn, Direction facing, BlockState facingState, LevelAccessor worldIn, BlockPos currentPos, BlockPos facingPos) {
         if (stateIn.getValue(WATERLOGGED)) {
             worldIn.scheduleTick(currentPos, Fluids.WATER, Fluids.WATER.getTickDelay(worldIn));
         }
-        if (!stateIn.canSurvive(worldIn, currentPos)) {
-            if (worldIn instanceof ServerLevel serverLevel) {
-                this.tick(stateIn, serverLevel, currentPos, worldIn.getRandom());
-            }
-            return stateIn;
-        }
+        worldIn.scheduleTick(currentPos, this, this.getDelayAfterPlace());
         return super.updateShape(stateIn, facing, facingState, worldIn, currentPos, facingPos);
     }